requirements.txt
python-dotenv

pydantic
pydantic-settings

pinecone

langchain
langchain-pinecone
langchain-openai
langchain-huggingface

langserve
langgraph

sse_starlette

fastapi
uvicorn

streamlit

